[change] copyToClipboard util 함수로 변경
